+++ ETL_Airflow/dags/tasks/m_customer_sales_report_task.py
@@ -1,6 +1,6 @@
 from airflow.decorators import task
 from airflow.exceptions import AirflowException
-from utils import init_spark, load_to_postgres, DuplicateValidator, fetch_env_schema,read_from_postgres
+from tasks.utils import init_spark, load_to_postgres, DuplicateValidator, fetch_env_schema,read_from_postgres
 import logging
 from pyspark.sql.functions import col, sum, current_date, round, when, date_sub, month, year, percent_rank, current_timestamp,lit,date_format
 from pyspark.sql.window import Window
