+++ ETL_Airflow/dags/tasks/m_product_performance_task.py
@@ -1,6 +1,6 @@
 from airflow.decorators import task
 from airflow.exceptions import AirflowException
-from utils import init_spark, load_to_postgres, DuplicateValidator, read_from_postgres, fetch_env_schema
+from tasks.utils import init_spark, load_to_postgres, DuplicateValidator, read_from_postgres, fetch_env_schema
 import logging
 from pyspark.sql.functions import col, sum, current_date, round, when
