+++ ETL_Airflow/dags/tasks/m_supplier_performance_task.py
@@ -1,6 +1,6 @@
 from airflow.decorators import task
 from airflow.exceptions import AirflowException
-from utils import init_spark, load_to_postgres, DuplicateValidator, fetch_env_schema,read_from_postgres
+from tasks.utils import init_spark, load_to_postgres, DuplicateValidator, fetch_env_schema,read_from_postgres
 import logging
 from pyspark.sql.functions import col, sum ,round, current_date, row_number, when,count,trim
 from pyspark.sql.window import Window
